+++ commands/breadbox/armCommands.py
@@ -1,8 +1,8 @@
 import commands2
 import commands2.button
 import commands2.cmd
-from subsystems.actuators.dumboArmController import ArmController
-from subsystems.actuators.dumboArmRotation import ArmRotation
+from subsystems.actuators.breadboxArmController import ArmController
+from subsystems.actuators.breadboxArmRotation import ArmRotation
 
 
 def createArmPositionCommands(controller: commands2.button.CommandGenericHID, armController: ArmController, arm_subsystem: ArmRotation):
